pyproject.toml
[tool.poetry]
name = "nonebot-plugin-giyf"
version = "0.1.0"
description = "nonebot2 群聊快速生成搜索引擎链接"
authors = ["KoishiChan <68314080+KoishiStudio@users.noreply.github.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/KoishiStudio/nonebot-plugin-giyf"

[tool.poetry.dependencies]
python = ">=3.7"
nonebot2 = "~=2.0.0b2"
nonebot-adapter-onebot = "~=2.0.0b1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
